Dynamic data from AngularJS query links to Chart page
test case was the company irib.  Clicking the lower button on this page pulls topics irib is interested in and write them to the page along with links.

The idea is to use this type of functionality to update the charts live, or at least to let the user browse topics on the same page as the chart for adding additional topics.